refactor(client/home/FeaturedSegment): encapsualmiento
